refactor: replace project list to course detail page
